second commit. append connecting to database and set settings to the database columns

+++ WebApplication/WebApplication/Models/ViewModels/MalfunctionViewModel.cs
@@ -27,6 +27,6 @@ namespace WebApplication.Models.ViewModels
         }
 
         // стоимость неисправности
-        public int Price { get; private set; }
+        public decimal Price { get; private set; }
     }
 }
